requirements.txt
flask
flask-cors
flask-limiter
python-dotenv
google-generativeai
google-cloud-aiplatform
vertexai
openai
matplotlib
numpy
networkx
Pillow
redis
gunicorn
pytest